refactor(sdk): split client.ts into focused modules
Split the monolithic client.ts (703 lines) into smaller, focused files:
- http.ts: HTTP utilities (safeFetch, error handling)
- text.ts: generateText function
- stream/sse.ts: SSE parser
- stream/index.ts: streamText function
- object.ts: generateObject function
- client.ts: Factory and re-exports
